AppADay 03
Created AppADay 03_scatter, an interactive scatter plot built in R Shiny that uses the built-in iris dataset. This app is relatively simple but (hopefully) will work better on mobile than my other apps. It is completed minus the hosting because I'm currently trying to find hosting alternatives to shinyapps.io